+++ libs/data_structures/matrix/matrix.h
@@ -40,7 +40,7 @@ void swapColumns(matrix *m, int j1, int j2);
 
 void insertionSortMatrixByCriteria(matrix *m, int (criteria)(int [], int), bool rowsOrCols);
 
-void insertionSortMatrixByCriteriaF(matrix *m, float (criteria)(int [], int), bool rowsOrCols);
+void insertionSortMatrixByCriteriaF(matrix *m, double (criteria)(int [], int), bool rowsOrCols);
 
 bool isSquareMatrix(matrix m);
